Extract creation compute into framework-free createKaraoke(); panel is now a thin caller
